- Changed: Additional memory reductions for game livings and some changes to the previous commit.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk@2412 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/commands/gmcommands/mob.cs
@@ -2251,12 +2251,12 @@ namespace DOL.GS.Commands
 
 		private void house(GameClient client, GameNPC targetMob, string[] args)
 		{
-			int house;
+			ushort house;
 			House H;
 
 			try
 			{
-				house = Convert.ToInt32(args[2]);
+				house = Convert.ToUInt16(args[2]);
 				H = HouseMgr.GetHouse(house);
 
 				if (H != null)
@@ -2349,9 +2349,9 @@ namespace DOL.GS.Commands
 
 		private void bodytype(GameClient client, GameNPC targetMob, string[] args)
 		{
-			int type;
+			ushort type;
 
-			if (args.Length > 2 && int.TryParse(args[2], out type))
+			if (args.Length > 2 && ushort.TryParse(args[2], out type))
 			{
 				targetMob.BodyType = type;
 				targetMob.SaveIntoDatabase();
